rename module and id
