+++ vortex-tensor/src/lib.rs
@@ -21,7 +21,7 @@ use crate::scalar_fns::inner_product::InnerProduct;
 use crate::scalar_fns::l2_denorm::L2Denorm;
 use crate::scalar_fns::l2_norm::L2Norm;
 use crate::scalar_fns::sorf_transform::SorfTransform;
-use crate::types::fixed_shape::FixedShapeTensor;
+use crate::types::fixed_shape_tensor::FixedShapeTensor;
 use crate::types::vector::Vector;
 
 pub mod matcher;
@@ -29,7 +29,7 @@ pub mod scalar_fns;
 
 mod types;
 
-pub use types::fixed_shape;
+pub use types::fixed_shape_tensor;
 pub use types::vector;
 
 pub mod encodings;
