+++ vortex-tensor/src/types/fixed_shape_tensor/metadata.rs
