+++ vortex-tensor/src/types/fixed_shape_tensor/vtable.rs
@@ -11,9 +11,9 @@ use vortex_error::vortex_bail;
 use vortex_error::vortex_ensure;
 use vortex_error::vortex_ensure_eq;
 
-use crate::types::fixed_shape::FixedShapeTensor;
-use crate::types::fixed_shape::FixedShapeTensorMetadata;
-use crate::types::fixed_shape::proto;
+use crate::types::fixed_shape_tensor::FixedShapeTensor;
+use crate::types::fixed_shape_tensor::FixedShapeTensorMetadata;
+use crate::types::fixed_shape_tensor::proto;
 
 impl ExtVTable for FixedShapeTensor {
     type Metadata = FixedShapeTensorMetadata;
@@ -22,7 +22,7 @@ impl ExtVTable for FixedShapeTensor {
     type NativeValue<'a> = &'a ScalarValue;
 
     fn id(&self) -> ExtId {
-        ExtId::new("vortex.fixed_shape_tensor")
+        ExtId::new("vortex.tensor.fixed_shape_tensor")
     }
 
     fn serialize_metadata(&self, metadata: &Self::Metadata) -> VortexResult<Vec<u8>> {
@@ -102,8 +102,8 @@ mod tests {
     use vortex_array::dtype::extension::ExtVTable;
     use vortex_error::VortexResult;
 
-    use crate::types::fixed_shape::FixedShapeTensor;
-    use crate::types::fixed_shape::FixedShapeTensorMetadata;
+    use crate::types::fixed_shape_tensor::FixedShapeTensor;
+    use crate::types::fixed_shape_tensor::FixedShapeTensorMetadata;
 
     /// Serializes and deserializes the given metadata through protobuf, asserting equality.
     fn assert_roundtrip(metadata: &FixedShapeTensorMetadata) -> VortexResult<()> {
