+++ vortex-tensor/src/utils.rs
@@ -332,8 +332,8 @@ pub mod test_helpers {
     use vortex_error::VortexResult;
 
     use crate::scalar_fns::l2_denorm::L2Denorm;
-    use crate::types::fixed_shape::FixedShapeTensor;
-    use crate::types::fixed_shape::FixedShapeTensorMetadata;
+    use crate::types::fixed_shape_tensor::FixedShapeTensor;
+    use crate::types::fixed_shape_tensor::FixedShapeTensorMetadata;
     use crate::types::vector::Vector;
 
     /// Builds a `FixedSizeList<T, list_size>` storage array from flat `elements`. The row count is
